chat - move setup into `chatSetup` and split up
