+++ src/vs/workbench/contrib/chat/browser/chatSetup/media/chatSetup.css
